AirFlower - Basic Implementation
AirFlowers can be placed and rotated in the scene as wished.

Then the flowers have to be added to a so called AirFlowerGroup, which
contains a TArray for AirFlowers.

The AirFlowerGroup locates and resizes itself at the begin of the game,
so it is not important, where it is located, but it is recommended to
subordinate the flowers to the group, after they have been placed and
assigned to the TArray.

After this, the group can be moved as a unit.

The SkeletalMesh (air_flower.uasset) has to be replaced by the flowers
mesh.