AirFlowComponent
A generic component which can be added to each actor.

It generates an area, applying a force to the
CharacterMovementComponent, in case the character is within a cone
starting at the pivot of the actor in the direction of it.

The applied force is adjustable by the distance, angle and strength and
interpolates between the center and the edges of the cone.