First Death Implementation
The character dies when he reaches a specific z-value. This value can be
adjusted so the character doesnt die in case, it isnt wanted.

After death the camera is detached from the character and a vignette
grows bigger and bigger so the screen goes black.

Minor adjustment of the air control are made as well...